requirements.txt
pydantic-ai[ollama]
pydantic
pydantic-graph
python-dotenv
logfire
httpx
aiohttp
pydantic-evals